Refactoring - move logic for generating Liquibase Change Sets into its own class
