+++ spring-data-relational/src/main/java/org/springframework/data/relational/core/mapping/schemasqlgeneration/SchemaSQLGenerationDataModel.java
@@ -157,159 +157,4 @@ public class SchemaSQLGenerationDataModel
     public List<TableModel> getTableData() {
         return tableData;
     }
-
-    public void generateLiquibaseChangeset(Database database, String changeLogFilePath) throws InvalidExampleException, DatabaseException, IOException, ChangeLogParseException {
-
-        String changeSetId = Long.toString(System.currentTimeMillis());
-        generateLiquibaseChangeset(database,changeLogFilePath, changeSetId, "Spring Data JDBC");
-    }
-
-    public void generateLiquibaseChangeset(Database database, String changeLogFilePath, String changeSetId, String changeSetAuthor) throws InvalidExampleException, DatabaseException, IOException, ChangeLogParseException {
-
-        CatalogAndSchema[] schemas = new CatalogAndSchema[] { database.getDefaultSchema() };
-        SnapshotControl snapshotControl = new SnapshotControl(database);
-
-        DatabaseSnapshot snapshot = SnapshotGeneratorFactory.getInstance().createSnapshot(schemas, database, snapshotControl);
-        Set<liquibase.structure.core.Table> tables = snapshot.get(liquibase.structure.core.Table.class);
-
-        SchemaSQLGenerationDataModel liquibaseModel = new SchemaSQLGenerationDataModel();
-
-        for (TableModel t : tableData) {
-            if (t.getSchema() == null || t.getSchema().isEmpty()) {
-                t.setSchema(database.getDefaultSchema().getCatalogName());
-            }
-        }
-
-        for (liquibase.structure.core.Table table : tables) {
-
-            // Exclude internal Liquibase tables from comparison
-            if (table.getName().startsWith("DATABASECHANGELOG")) {
-                continue;
-            }
-
-            SqlIdentifier tableName = new DerivedSqlIdentifier(table.getName(), true);
-            TableModel tableModel = new TableModel(table.getSchema().getCatalogName(), tableName);
-            liquibaseModel.getTableData().add(tableModel);
-
-            List<liquibase.structure.core.Column> columns = table.getColumns();
-            for (liquibase.structure.core.Column column : columns) {
-                SqlIdentifier columnName = new DerivedSqlIdentifier(column.getName(), true);
-                String type = column.getType().toString();
-                boolean nullable = column.isNullable();
-                ColumnModel columnModel = new ColumnModel(columnName, type, nullable, false);
-                tableModel.getColumns().add(columnModel);
-            }
-        }
-
-        SchemaDiff difference = diffModel(liquibaseModel);
-
-        File changeLogFile = new File(changeLogFilePath);
-
-        DatabaseChangeLog databaseChangeLog;
-
-        try {
-            YamlChangeLogParser parser = new YamlChangeLogParser();
-            DirectoryResourceAccessor resourceAccessor = new DirectoryResourceAccessor(changeLogFile.getParentFile());
-            ChangeLogParameters parameters = new ChangeLogParameters();
-            databaseChangeLog = parser.parse(changeLogFilePath, parameters, resourceAccessor);
-        } catch (Exception ex) {
-            databaseChangeLog = new DatabaseChangeLog(changeLogFilePath);
-        }
-
-        ChangeLogSerializer serializer = new YamlChangeLogSerializer();
-        ChangeSet changeSet = new ChangeSet(changeSetId, changeSetAuthor, false, false, "", "", "" , databaseChangeLog);
-
-        for (TableModel t : difference.getTableAdditions()) {
-            CreateTableChange newTable = createAddTableChange(t);
-            changeSet.addChange(newTable);
-        }
-
-        for (TableModel t : difference.getTableDeletions()) {
-            DropTableChange dropTable = createDropTableChange(t);
-            changeSet.addChange(dropTable);
-        }
-
-        for (TableDiff t : difference.getTableDiff()) {
-
-            if (t.getAddedColumns().size() > 0) {
-                AddColumnChange addColumnChange = new AddColumnChange();
-                addColumnChange.setSchemaName(t.getTableModel().getSchema());
-                addColumnChange.setTableName(t.getTableModel().getName().getReference());
-
-                for (ColumnModel column : t.getAddedColumns()) {
-                    AddColumnConfig addColumn = createAddColumnChange(column);
-                    addColumnChange.addColumn(addColumn);
-                }
-
-                changeSet.addChange(addColumnChange);
-            }
-
-            if (t.getDeletedColumns().size() > 0) {
-                DropColumnChange dropColumnChange = new DropColumnChange();
-                dropColumnChange.setSchemaName(t.getTableModel().getSchema());
-                dropColumnChange.setTableName(t.getTableModel().getName().getReference());
-
-                List<ColumnConfig> dropColumns = new ArrayList<ColumnConfig>();
-                for (ColumnModel column : t.getDeletedColumns()) {
-                    ColumnConfig config = new ColumnConfig();
-                    config.setName(column.getName().getReference());
-                    dropColumns.add(config);
-                }
-                dropColumnChange.setColumns(dropColumns);
-                changeSet.addChange(dropColumnChange);
-            }
-        }
-
-        List changes = new ArrayList<ChangeLogChild>();
-        for (ChangeSet change : databaseChangeLog.getChangeSets()) {
-            changes.add(change);
-        }
-        changes.add(changeSet);
-        FileOutputStream fos = new FileOutputStream(changeLogFile);
-        serializer.write(changes, fos);
-    }
-
-    private AddColumnConfig createAddColumnChange(ColumnModel column) {
-
-        AddColumnConfig config = new AddColumnConfig();
-        config.setName(column.getName().getReference());
-        config.setType(column.getType());
-
-        if (column.isIdentityColumn()) {
-            config.setAutoIncrement(true);
-        }
-        return config;
-    }
-
-    CreateTableChange createAddTableChange(TableModel table) {
-
-        CreateTableChange change = new CreateTableChange();
-        change.setSchemaName(table.getSchema());
-        change.setTableName(table.getName().getReference());
-
-        for (ColumnModel column : table.getColumns()) {
-            ColumnConfig columnConfig = new ColumnConfig();
-            columnConfig.setName(column.getName().getReference());
-            columnConfig.setType(column.getType());
-
-            if (column.isIdentityColumn()) {
-                columnConfig.setAutoIncrement(true);
-                ConstraintsConfig constraints =  new ConstraintsConfig();
-                constraints.setPrimaryKey(true);
-                columnConfig.setConstraints(constraints);
-            }
-            change.addColumn(columnConfig);
-        }
-
-        return change;
-    }
-
-    DropTableChange createDropTableChange(TableModel table) {
-        DropTableChange change = new DropTableChange();
-        change.setSchemaName(table.getSchema());
-        change.setTableName(table.getName().getReference());
-        change.setCascadeConstraints(true);
-
-        return change;
-    }
 }
